Sprint 2 -> Main  (#30)
* Added enums
Modified ID from Long to Integer

* changed the properties
Fixed schema

* updated pom

* updated pom with jacoco plugin (#19)


* created rule classes and service, txnDAO and repository, AlertRuleDAO and repository (#20)

corrected types in sql files


* fixed watchlist stored_proc and its tests (#22)


* tests created (#23)

fix for RuleEngineService


* Repositories for all tables created (#24)


* Feature/adding alert dao (#25)

* Add AlertRepository and AlertDao

* Implement AlertDAO and update AlertRepository

* Feature/implement watchlist and transaction (#27)

* Implement WatchlistDao

* Implement WatchlistMatchDao and WatchlistMatchRepository

* Add upddate and get status methods to TxnRepository and TxnDAO

* Update txn status options and implement WatchlistScreenignService

* Stage txn status options

* Refactore of the unit tests (#26)

Created integration test
Added test profile
Introduced H2 database for testing


* Fix WatchlistMatchRepository query method and test context loading (#28)

Rename findByTxnIdAndMatchScoreGreaterThanEqual to
findByTxn_TxnIdAndMatchScoreGreaterThanEqual to correctly traverse
the txn association — Spring Data was resolving 'TxnId' as txn.id
which does not exist on the Txn entity.

Update WatchlistMatchDao call site to match.


* Updated README.md file (#29)


---------

+++ src/main/java/com/grad/sam/model/Alert.java
@@ -20,7 +20,7 @@ public class Alert {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "alert_id")
-    private Long alertId;
+    private Integer alertId;
 
     @Column(name = "alert_ref", nullable = false, unique = true, length = 15)
     private String alertRef;
